perf(round-3a): tighten SymbolNode integer widths to u32
Three fields on `SymbolNode` were `usize` (8 bytes each on 64-bit) when
`u32` is plenty:
  - `byte_range: Range<usize>` — a source file fits in 4 GB
  - `line_range: Range<usize>` — line counts fit in u32 by many orders
  - `estimated_tokens: usize`  — same

Switch all three to `u32` (Range<u32> for the ranges). Saves 8 B × 2
range fields + 4 B = 20 B/symbol nominally; measured 24 B/symbol after
factoring in struct alignment changes. `AgentToolCall.target_lines` also
moves to `Option<Range<u32>>` so the symbol/line-overlap comparison in
`app::symbol_lines_match` stays a clean integer comparison without
casts.

Conversion happens at construction sites in each parser
(`byte_range.start as u32..byte_range.end as u32`) and at slice sites
(`as usize` to feed into `&src[..]`); tree-sitter still returns
`Range<usize>` so the boundary lives there. A few `ledger.record()`
call sites pass `sym.estimated_tokens as usize` since the ledger's
token counter remains `usize` (it sums across many symbols).

Measured impact (heap_probe, dhat, same machine):

  tier   peak bytes (R2 → R3a)              peak blocks (unchanged)
  10k    2,570,839 → 2,330,839     (-9.3%)  14,834
  100k  25,691,839 → 23,291,839    (-9.3%) 148,034
  500k 128,451,839 → 116,451,839   (-9.3%) 740,034

  → 12 MB saved at the 500k tier with zero block-count change
    (these are pure per-symbol-size reductions, not allocation count)
  → per-symbol cost: 257 B → 233 B average

Time benches (`symbol_pipeline`) all within noise — type changes don't
affect the work being measured.

Cumulative effect of Round 0 → Round 3a on 500k symbols:
  peak heap:    155.0 MB → 116.5 MB   (-24.9%)
  blocks at peak: 1.61M → 740k        (-54%)
  merkle pass:   57.13 ms → 26.53 ms  (-54%)

+++ benches/support/fixtures.rs
@@ -1089,7 +1089,7 @@ pub fn make_populated_ledger(project: &ProjectTree) -> ContextLedger {
     let hash = [0u8; 32];
     for file in &project.files {
         for sym in &file.symbols {
-            ledger.record(sym.id.clone(), ReadDepth::FullBody, hash, "agent-0".to_string(), sym.estimated_tokens);
+            ledger.record(sym.id.clone(), ReadDepth::FullBody, hash, "agent-0".to_string(), sym.estimated_tokens as usize);
         }
     }
     ledger
