+++ src/serena/mod.rs
@@ -212,12 +212,12 @@ fn convert_symbol(
         category,
         label,
         file_path: std::sync::Arc::clone(file_path),
-        byte_range: (start_line * 40 + start_char)..(end_line * 40 + end_char),
-        line_range: (start_line + 1)..(end_line + 1), // 1-indexed like tree-sitter
+        byte_range: ((start_line * 40 + start_char) as u32)..((end_line * 40 + end_char) as u32),
+        line_range: ((start_line + 1) as u32)..((end_line + 1) as u32), // 1-indexed like tree-sitter
         content_hash,
         merkle_hash: [0u8; 32],
         children,
-        estimated_tokens: line_count * 15,
+        estimated_tokens: (line_count * 15) as u32,
     };
     compute_merkle_hash(&mut node);
     Ok(node)
@@ -246,7 +246,7 @@ fn extract_range(val: &Value) -> (usize, usize, usize, usize) {
 fn estimate_total_lines(symbols: &[SymbolNode]) -> usize {
     symbols
         .iter()
-        .map(|s| s.line_range.end)
+        .map(|s| s.line_range.end as usize)
         .max()
         .unwrap_or(0)
 }
